# Optimize sparse 2:4 compression performance (3.69x speedup)