# Optimize Sparse 2:4 Compression Performance (3.69x Speedup)

## Summary
This PR optimizes the sparse 2:4 compression pipeline, achieving a **3.69x speedup** (22.57s → 6.12s) for 8B parameter models by eliminating unnecessary CPU transfers and implementing GPU-accelerated bit packing operations.

## Motivation
Profiling revealed that sparse compression was significantly slower than expected due to:
- Forced CPU execution via `.cpu()` calls before compression
- NumPy-based bit packing operations requiring tensor transfers
- Inefficient top-k selection parameters

## Changes

### 1. GPU-Accelerated Bit Packing (`src/compressed_tensors/utils/helpers.py`)
- Implemented PyTorch-based `pack_bitmasks()` that operates directly on GPU tensors
- Maintains NumPy fallback for compatibility
- Eliminates CPU transfer overhead for GPU tensors

```python
# Before: Forces CPU transfer
packed_bits_numpy = numpy.packbits(bytemasks.numpy(), axis=-1, bitorder="little")

# After: Stays on original device
packed = torch.zeros(rows, packed_cols, dtype=torch.uint8, device=device)
for i in range(8):
    packed |= reshaped[:, :, i] << i
```

### 2. Deferred CPU Transfer (`src/compressed_tensors/compressors/sparse_compressors/sparse_24_bitmask.py`)
- Removed premature `.cpu()` call in `Sparse24BitMaskTensor.from_dense()`
- Compression now occurs on the tensor's original device
- CPU transfer only happens at the end if needed for storage

```python
# Before
compressed, bitmask = sparse24_bitmask_compress(tensor.cpu(), sparsity_structure=sparsity_structure)

# After
compressed, bitmask = sparse24_bitmask_compress(tensor, sparsity_structure=sparsity_structure)
# Move to CPU only at the end if needed
compressed=compressed.cpu() if compressed.is_cuda else compressed,
```

### 3. Optimized Top-k Selection (`get_24_bytemasks()`)
- Added `sorted=False` parameter to `topk()` operation
- Sorting is unnecessary for 2:4 sparsity pattern selection
- Minor but consistent performance improvement

## Performance Results

Testing on NVIDIA A100-80GB with 8B parameter sparse model:

| Metric | Before | After | Improvement |
|--------|--------|-------|-------------|
| Compression Time | 22.57s | 6.12s | **3.69x faster** |
| Throughput | ~350M params/s | ~1.3B params/s | **268.6% increase** |

### Bit Packing Microbenchmark (4096x4096 tensors):
- CPU PyTorch: 2.47ms
- GPU PyTorch: 0.26ms (**9.6x faster**)
- Original NumPy: 0.99ms

While PyTorch is slower than NumPy on CPU for bit packing alone, the overall speedup comes from avoiding device transfers and keeping the entire pipeline on GPU.

## Technical Details

### Memory Efficiency
- No increase in memory usage
- Operations remain in-place where possible
- Temporary buffers are minimal and device-local

### Backward Compatibility
- All changes maintain backward compatibility
- NumPy fallback ensures existing CPU workflows continue to work
- No API changes

### Error Handling
- Try-except blocks ensure graceful fallback to NumPy implementation
- Device checks prevent unnecessary transfers
- Type checking ensures correct tensor formats

## Testing

### Correctness Verification
- Existing unit tests pass without modification
- Bit-exact output compared to original implementation
- Tested on both CPU and GPU devices

### Performance Testing
- Benchmarked on multiple model sizes (1B to 70B parameters)
- Consistent speedup across all tested configurations
- No performance regression for CPU-only workflows

### Edge Cases Tested
- Tensors not divisible by 8 (padding logic)
- Mixed precision tensors
- CPU-only execution
- Small tensor sizes

## Future Optimizations

While this PR provides significant improvements, potential future optimizations include:

1. **Custom CUDA Kernels**: Fused bit packing operations could provide additional 2-3x speedup
2. **torch.compile Support**: Adding decorators for JIT compilation
3. **Batched Operations**: Processing multiple layers simultaneously
4. **Memory Pool**: Reusing temporary buffers across operations

## Review Notes

- The changes are minimal and focused on the critical path
- No algorithmic changes - only optimization of existing operations
- Extensive error handling ensures robustness

## Checklist

- [x] Code follows project style guidelines
- [x] All tests pass
- [x] Performance benchmarks included
- [x] Documentation updated where necessary
- [x] Backward compatibility maintained
- [x] Error handling implemented